+++ packages/contracts/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Created the `IMultisig` interface.
+- Added `PlaceholderSetup` contract and adapted deploy scripts.
 
 ### Changed